Update to PHP8.1
The container is already running on PHP8.1, but the test containers are still running 7.4, so update them.

I had to update some tests to provide valid data, because PHP8.1 doesn't like referencing undefined array keys. As part of this I updated SqsAdapter to check that `sendby.documentType` exists before accessing it.

Update the target platform in composer.json.

Add `composer check-platform-reqs` to the Dockerfile to ensure that the container can support the dependencies' requirements.

#patch